+++ src/routes/_components/ArticleList/ArticlePreview.html
@@ -12,9 +12,9 @@
 			</span>
 		</div>
 
-		{#if $user}
+		{#if user}
 			<div class="pull-xs-right">
-				<button class='btn btn-sm {article.favorited ? "btn-primary" : "btn-outline-primary"}' on:click='toggleFavorite()'>
+				<button class='btn btn-sm {article.favorited ? "btn-primary" : "btn-outline-primary"}' on:click='{toggleFavorite}'>
 					<i class="ion-heart"></i> {article.favoritesCount}
 				</button>
 			</div>
@@ -38,12 +38,12 @@
 <script>
 	import * as api from '../../_api.js';
 
-	export let article;
+	export let article, user;
 
 	function toggleFavorite() {
 		const promise = article.favorited ?
-			api.del(`articles/${article.slug}/favorite`, $user && $user.token) :
-			api.post(`articles/${article.slug}/favorite`, null, $user && $user.token);
+			api.del(`articles/${article.slug}/favorite`, user && user.token) :
+			api.post(`articles/${article.slug}/favorite`, null, user && user.token);
 
 		promise.then(response => {
 			this.set(response);
