+++ src/routes/_components/ArticleList/index.html
@@ -6,10 +6,10 @@
 	{:else}
 		<div>
 			{#each articles as article (article.slug)}
-				<ArticlePreview {article}/>
+				<ArticlePreview {article} user={$session.user}/>
 			{/each}
 
-			<ListPagination {articlesCount} {page} on:select='{({ detail }) => page = detail.page })}' />
+			<ListPagination {articlesCount} {page} on:select='{ ({ detail }) => page = detail.page }' />
 		</div>
 	{/if}
 {:else}
@@ -17,13 +17,15 @@
 {/if}
 
 <script>
+	import { stores } from '@sapper/app';
 	import ArticlePreview from './ArticlePreview.html';
 	import ListPagination from './ListPagination.html';
 	import * as api from '../../_api.js';
 
-	export let tab, username, favourites, tag;
+	export let tab, username, favorites, tag;
+	const { session } = stores();
 
-	let page = 0, query, page;
+	let page = 0, query, articles, articlesCount;
 	
 	$: {
 		const endpoint = tab === 'feed' ? 'articles/feed' : 'articles';
@@ -41,14 +43,16 @@
 	function getData() {
 		articles = null;
 
-		api.get(current.query, $user && $user.token)
+		api.get(query, $session.user && $session.user.token)
 			.then(data => {
 				console.log(data)
 				// TODO do we need some error handling here?
 			//	this.set(data);
+				articles = data.articles;
+				articlesCount = data.articlesCount
 			});
 	}
-
+	$: console.log(articles)
 		// computed: {
 		// 	 query: ({ tab, page, username, favorites, tag }) => {
 		// 		const endpoint = tab === 'feed' ? 'articles/feed' : 'articles';
