+++ src/routes/_components/MainView/index.html
@@ -1,7 +1,7 @@
 <div class="col-md-9">
 	<div class="feed-toggle">
 		<ul class="nav nav-pills outline-active">
-			{#if $user}
+			{#if $session.user}
 				<li class="nav-item">
 					<a href="." class='nav-link {tab === "feed" ? "active" : "" }' on:click='{yourFeed}'>
 						Your Feed
@@ -35,9 +35,11 @@
 </div>
 
 <script>
+	import { stores } from '@sapper/app';
 	import ArticleList from '../ArticleList/index.html';
 
 	export let tab = 'all', tag;
+	const { session } = stores();
 
 	function yourFeed() {
 		tab = "feed";
