+++ src/routes/_components/Nav.html
@@ -3,37 +3,37 @@
 		<a rel='prefetch' class="navbar-brand" href=".">conduit</a>
 		<ul class="nav navbar-nav pull-xs-right">
 			<li class="nav-item">
-				<a rel='prefetch' class="nav-link {path === '/' ? 'active' : ''}" href="/">Home</a>
+				<a rel='prefetch' class="nav-link {$page.path === '/' ? 'active' : ''}" href="/">Home</a>
 			</li>
 
-			{#if $user}
+			{#if $session.user}
 				<li class="nav-item">
-					<a rel='prefetch' href="/editor" class="nav-link {path === '/editor' ? 'active' : ''}">
+					<a rel='prefetch' href="/editor" class="nav-link {$page.path === '/editor' ? 'active' : ''}">
 						<i class="ion-compose"></i>&nbsp;New Post
 					</a>
 				</li>
 
 				<li class="nav-item">
-					<a rel='prefetch' href="/settings" class="nav-link {path === '/settings' ? 'active' : ''}">
+					<a rel='prefetch' href="/settings" class="nav-link {$page.path === '/settings' ? 'active' : ''}">
 						<i class="ion-gear-a"></i>&nbsp;Settings
 					</a>
 				</li>
 
 				<li class="nav-item">
-					<a rel='prefetch' href='/@{$user.username}' class="nav-link">
+					<a rel='prefetch' href='/@{$session.user.username}' class="nav-link">
 						<!-- <img src={$user.image} class="user-pic" alt={$user.username}> -->
-						{$user.username}
+						{$session.user.username}
 					</a>
 				</li>
 			{:else}
 				<li class="nav-item">
-					<a rel='prefetch' href="/login" class="nav-link {path === '/login' ? 'active' : ''}">
+					<a rel='prefetch' href="/login" class="nav-link {$page.path === '/login' ? 'active' : ''}">
 						Sign in
 					</a>
 				</li>
 
 				<li class="nav-item">
-					<a rel='prefetch' href="/register" class="nav-link {path === '/register' ? 'active' : ''}">
+					<a rel='prefetch' href="/register" class="nav-link {$page.path === '/register' ? 'active' : ''}">
 						Sign up
 					</a>
 				</li>
@@ -43,6 +43,7 @@
 </nav>
 
 <script>
-	// path store?
-	export let path;
+	import { stores } from '@sapper/app';
+
+	const { page, session } = stores();
 </script>
