+++ src/routes/_components/Tags.html
@@ -1,7 +1,7 @@
 {#if tags}
 	<div className="tag-list">
 		{#each tags as tag}
-			<a href="." class="tag-default tag-pill" on:click='dispatch("select", { tag })'>
+			<a href="." class="tag-default tag-pill" on:click='{() => dispatch("select", { tag })}'>
 				{tag}
 			</a>
 		{/each}
