+++ src/routes/_error.html
@@ -34,5 +34,7 @@
 </style>
 
 <script>
+  export let error, status
+
   let dev = process.env.NODE_ENV === "development";
 </script>
