+++ src/routes/login/index.html
@@ -33,13 +33,14 @@
 <script>
 	import { goto } from '../../../__sapper__/client.js';
 	import ListErrors from '../_components/ListErrors.html';
+	import { userSession } from '../../store.js';
 
 	let email = '',
 	password = '',
 	errors = null
 			
 	function submit(event) {
-		this.store.login({ email, password })
+		userSession.login({ email, password })
 			.then(response => {
 				if (response.errors) {
 					errors = response.errors;
