+++ src/routes/register/index.html
@@ -37,14 +37,15 @@
 <script>
 	import { goto } from '../../../__sapper__/client.js';
 	import ListErrors from '../_components/ListErrors.html';
-	
+	import { userSession } from '../../store.js';
+
 	let username = '',
 	email = '',
 	password = '',
 	errors = null;
 				
 	function submit(event) {
-		this.store.register({ username, email, password })
+		userSession.register({ username, email, password })
 			.then(response => {
 				if (response.errors) {
 					errors = response.errors;
