+++ src/service-worker.js
@@ -1,4 +1,4 @@
-import { files, shell, routes, timestamp } from '../__sapper__/service-worker.js';
+import { files, shell, routes, timestamp } from '@sapper/service-worker.js';
 
 const ASSETS = `cache${timestamp}`;
 
@@ -74,7 +74,7 @@ self.addEventListener('fetch', event => {
 					const response = await fetch(event.request);
 					cache.put(event.request, response.clone());
 					return response;
-				} catch(err) {
+				} catch (err) {
 					const response = await cache.match(event.request);
 					if (response) return response;
